+++ CHANGES.md
@@ -22,6 +22,7 @@ Release Notes.
 - Metadata: Wait for the existing registration to be removed before registering the node.
 - Stream: Introduce the batch scan to improve the performance of the query and limit the memory usage.
 - Add memory protector to protect the memory usage of the system. It will limit the memory usage of the querying.
+- Metadata: Introduce the periodic sync to sync the metadata from the etcd to the local cache in case of the loss of the events.
 
 ### Bug Fixes